Feature/portable toolchain (#6)
* remove hardcoded llvm path in cmake

* fix LLVM version at 23

* add lit testing to CI

* test on macos in CI, fix Dockerfile's LLVM version

* add missing include in LRTableSerializer

* add native components to llvm_libs in CMake

* update versions in GitHub Actions

* remove duplicate libraries bug in CMake

* update brew in CI

* fix ubuntu version in Dockerfile

* disable type storage test

* change int64_t to long long

When cpp compiler mangles names, int64_t ends up as either "x" or "l" depending on the OS. long long is mangled identically on both system.

* replace max-time with timeout in lit tests

* add a guard for `cd build` in testing script

* add hints for LLVM tools discovery

+++ examples/sources/print.cpp
@@ -13,7 +13,7 @@ struct Rectangle {
     height(height) {}
 };
 
-void print(int64_t value) {
+void print(long long value) {
   std::cout << value << std::endl;
 }
